refactor(web): move header tests into sibling __tests__ directories
