+++ web/app/components/header/account-setting/model-provider-page/__tests__/atoms.spec.tsx
@@ -7,7 +7,7 @@ import {
   useModelProviderListExpanded,
   useResetModelProviderListExpanded,
   useSetModelProviderListExpanded,
-} from './atoms'
+} from '../atoms'
 
 const createWrapper = () => {
   return ({ children }: { children: ReactNode }) => (
