+++ web/app/components/header/account-setting/model-provider-page/__tests__/supports-credits.spec.ts
@@ -1,6 +1,6 @@
-import type { ModelProvider } from './declarations'
-import { CurrentSystemQuotaTypeEnum } from './declarations'
-import { providerSupportsCredits } from './supports-credits'
+import type { ModelProvider } from '../declarations'
+import { CurrentSystemQuotaTypeEnum } from '../declarations'
+import { providerSupportsCredits } from '../supports-credits'
 
 vi.mock('@/config', async (importOriginal) => {
   const actual = await importOriginal<typeof import('@/config')>()
