+++ web/app/components/header/account-setting/model-provider-page/provider-added-card/__tests__/use-trial-credits.spec.ts
@@ -1,5 +1,5 @@
 import { renderHook } from '@testing-library/react'
-import { useTrialCredits } from './use-trial-credits'
+import { useTrialCredits } from '../use-trial-credits'
 
 const mockUseCurrentWorkspace = vi.fn()
